+++ internal/bundler_tests/bundler_tsconfig_test.go
@@ -2581,3 +2581,76 @@ func TestTsconfigJsonAsteriskNameCollisionIssue3354(t *testing.T) {
 		},
 	})
 }
+
+// https://github.com/evanw/esbuild/issues/3698
+func TestTsconfigPackageJsonExportsYarnPnP(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/packages/app/index.tsx": `
+				console.log(<div/>)
+			`,
+			"/Users/user/project/packages/app/tsconfig.json": `
+				{
+					"extends": "tsconfigs/config"
+				}
+			`,
+			"/Users/user/project/packages/tsconfigs/package.json": `
+				{
+					"exports": {
+						"./config": "./configs/tsconfig.json"
+					}
+				}
+			`,
+			"/Users/user/project/packages/tsconfigs/configs/tsconfig.json": `
+				{
+					"compilerOptions": {
+						"jsxFactory": "success"
+					}
+				}
+			`,
+			"/Users/user/project/.pnp.data.json": `
+				{
+					"packageRegistryData": [
+						[
+							"app",
+							[
+								[
+									"workspace:packages/app",
+									{
+										"packageLocation": "./packages/app/",
+										"packageDependencies": [
+											[
+												"tsconfigs",
+												"workspace:packages/tsconfigs"
+											]
+										],
+										"linkType": "SOFT"
+									}
+								]
+							]
+						],
+						[
+							"tsconfigs",
+							[
+								[
+									"workspace:packages/tsconfigs",
+									{
+										"packageLocation": "./packages/tsconfigs/",
+										"packageDependencies": [],
+										"linkType": "SOFT"
+									}
+								]
+							]
+						]
+					]
+				}
+			`,
+		},
+		entryPaths:    []string{"/Users/user/project/packages/app/index.tsx"},
+		absWorkingDir: "/Users/user/project",
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+	})
+}